pyproject.toml
[project]
name = "pykal"
version = "2.0.0"
description = "Personal HTML calendar generator"
requires-python = ">=3.11"
dependencies = [
    "ephem>=4.1",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
pykal = "pykal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykal"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
